fix(pacquet-cli): align list with pnpm flags, filtering, peers, and error handling
- Changed --production to --prod (matching pnpm CLI surface)
- lockfile_only now branches between current/wanted lockfile
- Selects correct workspace importer instead of always root_project()
- Filters params after full tree build, preserving transitive matches
- Walks optional_dependencies in snapshot traversal
- Skips peer DepNode push when --exclude-peers is set
- Propagates read_root_manifest errors instead of swallowing

---
Written by an agent (opencode, big-pickle).
